Closes #3781 move random module to numpy submodule
