+++ arkouda/numpy/random/__init__.py
@@ -1,0 +1,9 @@
+from ._generator import Generator, default_rng  # noqa
+from ._legacy import randint, standard_normal, uniform
+
+__all__ = [
+    'Generator',
+    'randint',
+    'standard_normal',
+    'uniform',
+]
