+++ arkouda/numpy/random/_legacy.py
