+++ arkouda/random/__init__.py
@@ -1,9 +1,3 @@
-from ._generator import Generator, default_rng  # noqa
-from ._legacy import randint, standard_normal, uniform
+# flake8: noqa
 
-__all__ = [
-    'Generator',
-    'randint',
-    'standard_normal',
-    'uniform',
-]
+from arkouda.numpy.random import *
