+++ pytest.ini
@@ -34,10 +34,10 @@ testpaths =
     tests/numpy/manipulation_functions_test.py
     tests/numpy/numeric_test.py
     tests/numpy/numpy_test.py
+    tests/numpy/random_test.py
     tests/operator_test.py
     tests/pdarray_creation_test.py
     tests/pdarrayclass_test.py
-    tests/random_test.py
     tests/regex_test.py
     tests/scipy/scipy_test.py
     tests/security_test.py
